mocks/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
PyJWT[crypto]==2.10.1
python-multipart==0.0.20
